requirements.txt
flask>=3.0
pillow>=10.0
gunicorn>=21.0
PyMuPDF>=1.24
EbookLib>=0.18
beautifulsoup4>=4.12
